Conf product may have conf. attribute not used in layered nav
When configurable attribute is not used in LayeredNav, L155 fails indexation due to attribute is not in array.

E.g. you have clothing with color and size, where color is always some complex naming "crimson/blood red/peach" and color_group for this is "red". Color is not used in layered while color_group is. Such solution results in Notice: Undefined offset: <colorAttributeId> which is fixable by isset proposed here.