Updated ClientConfiguration to use constants for XML configuration paths
This is to allow the same configuration paths to be accessed from contexts where StoreManager or other dependencies may not be initialised. For example, accessing the configuration path as part of Magento's setup:install command.

An alternative to this would be to access each piece of the configuration path via their prior constants, but this only avoids replicating the configuration node strings themselves and replicates the hierachy in another part of the codebase, so this seems like a more simple solution.

Removed extension of AbstractConfiguration class, as this is not currently needed by the class.