Changed ClientConfiguration instance used during ConfigOptionsList to ensure that ScopeConfigInteface isn't accessed prior to Magento initialization
The custom options array previously passed to Client\ClientConfiguration was to allow for the setup:install parameters supplied via CLI to be passed to the client in order to validate the connection settings before proceeding with Magento installation.

This has been removed, as it doesn't look like there is currently any use cases that require it. It has been extracted into a separate ClientConfigurationInterface implementation so that ConfigOptionsList can construct an instance without touching ScopeConfigInterface, as this caused the following:

- Notice: Undefined index: websites in module-config/App/Config/Type/System.php on line 248
- Empty config to be saved to cache prior to Magento install (unknown what this causes currently)

Fixes #738