Fix Attribute values not showing up
This is probably the wrong way of doing it, but it works.
I migrated from 1.9.2.2 to 2.2.6, and when I go to stores -> Attributes -> Product, and select a attribute with a dropdown type, I see no attributes in there. If I add them they show up in the front end, Cannot remove any (because they are not there)
Disabling Smile_ElasticsuiteCatalog fixes the issue.
Diffing the HTML output with Smile_ElasticsuiteCatalog enabled and disabled reveals that this code is missing.