[Core] Prevent stacking SQL where prod. id conditions
That can happen when a single collection is used in a loop with
different search queries each time.
While the Elasticsearch requests will be correct, the collection will
only be correctly hydrated the first time.
Especially visible with a pageSize of 1 and search terms generating
specific and non overlapping results sets.